requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
celery==5.4.0
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==42.0.5
daphne==4.1.2
deprecation==2.1.0
django-celery-beat==2.6.0
django-ckeditor-5==0.2.12
django-js-asset==2.2.0
django-phonenumber-field==7.3.0
django-phonenumbers==1.0.1
django-timezone-field==6.1.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
kombu==5.3.7
msgpack==1.0.8
mysql==0.0.3
mysqlclient==2.2.4
packaging==24.0
phonenumbers==8.13.33
pillow==10.2.0
prompt-toolkit==3.0.43
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.3
service-identity==24.1.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2024.1
validators==0.24.0
vine==5.1.0
wcwidth==0.2.13
zope.interface==6.3
